CSS Overhaul
Starting to overhaul the CSS on the project to focus on mobile-first development/design. I also fixed a bug in the Javascript where when submitting the function it would add a bunch of numbers after the decimal instead of keeping it at 100th.

Tested that function with several different number and it works correctly now!